+++ router-tests/jwks/jwks.go
@@ -1,6 +1,7 @@
 package jwks
 
 import (
+	"context"
 	"crypto/rand"
 	"crypto/rsa"
 	"encoding/base64"
@@ -8,6 +9,8 @@ import (
 	"math/big"
 	"net/http"
 	"net/http/httptest"
+	"testing"
+	"time"
 
 	"github.com/golang-jwt/jwt/v5"
 )
@@ -79,7 +82,23 @@ func (s *Server) JWKSURL() string {
 	return s.httpServer.URL + jwksHTTPPath
 }
 
-func NewServer() (*Server, error) {
+func (s *Server) waitForServer(ctx context.Context) error {
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		default:
+		}
+		resp, err := http.Get(s.httpServer.URL)
+		if err == nil {
+			_ = resp.Body.Close()
+			return nil
+		}
+		time.Sleep(time.Millisecond * 100)
+	}
+}
+
+func NewServer(t *testing.T) (*Server, error) {
 	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
 	if err != nil {
 		return nil, err
@@ -90,5 +109,10 @@ func NewServer() (*Server, error) {
 	mux := http.NewServeMux()
 	mux.HandleFunc(jwksHTTPPath, s.jwksJSON)
 	s.httpServer = httptest.NewServer(mux)
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
+	defer cancel()
+	if err := s.waitForServer(ctx); err != nil {
+		t.Fatal(err)
+	}
 	return s, nil
 }